backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0


sqlalchemy==2.0.32
psycopg[binary]==3.2.1
alembic==1.13.2

passlib==1.7.4
bcrypt==3.2.2
python-jose==3.3.0

redis==5.0.8


scikit-learn==1.5.2
joblib==1.4.2
numpy==2.0.2
pandas==2.2.2

minio>=7.2.0
python-multipart>=0.0.9
fpdf2>=2.7.9
apscheduler>=3.10.4
anthropic>=0.40.0